+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.5] - April 24, 2024
+
+- Add keyword argument `default_to_cl100k` to `count_tokens_for_message` function to allow for defaulting to the CL100k token limit if the model is not found.
+- Add keyword argument `default_to_minimum` to `get_token_limit` function to allow for defaulting to the minimum token limit if the model is not found.
+
 ## [0.0.4] - April 21, 2024
 
 - Rename to openai-messages-token-helper from llm-messages-token-helper to reflect library's current OpenAI focus.